+++ src/MenuState.cpp
@@ -1,5 +1,5 @@
 #include "MenuState.hpp"
 
 MenuState::MenuState() {}
-void MenuState::onUpdate(double dt) {}
+void MenuState::onUpdate(float dt) {}
 void MenuState::onDraw(sf::RenderTarget &target) {}
